+++ packages/@runlightyear/lightyear/src/base/auth.ts
@@ -17,6 +17,14 @@ export interface AuthProps {
  */
 export type AuthData = {
   /**
+   * The username for basic authentication
+   */
+  username: string | null;
+  /**
+   * The password for basic authentication
+   */
+  password: string | null;
+  /**
    * The api key
    */
   apiKey: string | null;
@@ -57,16 +65,19 @@ export async function getAuthData(name: string): Promise<AuthData> {
 
   const data = (await response.json()) as AuthData;
 
-  const { accessToken, refreshToken, apiKey } = data;
+  const { accessToken, refreshToken, apiKey, username, password } = data;
 
   accessToken && prefixedRedactedConsole.addSecrets([accessToken]);
   refreshToken && prefixedRedactedConsole.addSecrets([refreshToken]);
   apiKey && prefixedRedactedConsole.addSecrets([apiKey]);
+  apiKey && prefixedRedactedConsole.addSecrets([password]);
 
   return {
     accessToken,
     refreshToken,
     apiKey,
+    username,
+    password,
   };
 }