+++ packages/@runlightyear/lightyear/src/base/http.ts
@@ -8,7 +8,7 @@ export type HttpMethod = "get" | "post" | "put" | "patch" | "delete";
 /**
  * @public
  */
-export interface HttpProxyRequestOptions {
+export interface HttpProxyRequestProps {
   method?: HttpMethod;
   url: string;
   params?: Record<string, any>;
@@ -54,15 +54,15 @@ export class HttpProxyResponseError extends Error {
 }
 
 export interface HttpRequest {
-  (options: HttpProxyRequestOptions): Promise<HttpProxyResponse>;
+  (props: HttpProxyRequestProps): Promise<HttpProxyResponse>;
 }
 
-export const httpRequest: HttpRequest = async (options) => {
-  console.debug("httpRequest with options", JSON.stringify(options, null, 2));
+export const httpRequest: HttpRequest = async (props) => {
+  console.debug("httpRequest with props", JSON.stringify(props, null, 2));
 
   const response = await baseRequest({
     uri: "/api/v1/httpRequest",
-    data: options,
+    data: props,
   });
 
   console.debug(`response.status`, response.status);