+++ packages/@runlightyear/lightyear/src/connectors/OAuthConnector.ts
@@ -6,17 +6,17 @@ import inDevelopment from "../util/inDevelopment";
 /**
  * @public
  */
-export interface OauthConnectorProps {
+export interface OAuthConnectorProps {
   appName?: string;
   customAppName?: string;
-  oauthConfigData: OauthConfigData;
+  oauthConfigData: OAuthConfigData;
   authData: AuthData;
 }
 
 /**
  * @public
  */
-export interface OauthConfigData {
+export interface OAuthConfigData {
   clientId: string | null;
   clientSecret: string | null;
   authRequestUrl: string | null;
@@ -25,21 +25,21 @@ export interface OauthConfigData {
 /**
  * @public
  *
- * Oauth2 Connector
+ * OAuth2 Connector
  *
- * Creates the input to the 3 essential Oauth tasks:
+ * Creates the input to the 3 essential OAuth tasks:
  *
  * 1. Generates a url to request an access token
  * 2. Converts the oauth callback parameters into the https call to request an access token
  * 3. Provides the https params to refresh an access token
  */
-export abstract class OauthConnector {
+export abstract class OAuthConnector {
   appName?: string;
   customAppName?: string;
-  oauthConfigData: OauthConfigData;
+  oauthConfigData: OAuthConfigData;
   authData?: AuthData;
 
-  protected constructor(props: OauthConnectorProps) {
+  protected constructor(props: OAuthConnectorProps) {
     const { oauthConfigData, appName, customAppName, authData } = props;
     this.oauthConfigData = oauthConfigData;
 
@@ -166,7 +166,15 @@ export abstract class OauthConnector {
       expiresAt = dayjsUtc().add(parseInt(expiresIn), "seconds").format();
     }
 
-    return { tokenType, accessToken, refreshToken, expiresAt, apiKey: null };
+    return {
+      tokenType,
+      accessToken,
+      refreshToken,
+      expiresAt,
+      apiKey: null,
+      username: null,
+      password: null,
+    };
   }
 
   getRefreshAccessTokenHeaders() {