+++ packages/@runlightyear/openai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @runlightyear/openai
 
+## 0.2.0
+
+### Minor Changes
+
+- 6fb536f: Updated documentation
+
+### Patch Changes
+
+- Updated dependencies [6fb536f]
+- Updated dependencies
+- Updated dependencies [6fb536f]
+- Updated dependencies [6fb536f]
+  - @runlightyear/lightyear@0.17.0
+
 ## 0.1.2
 
 ### Patch Changes